feat(text): split layouts and preserve whitespace interaction
